Prep release/ directory with some basic docker work.
1. settings.yml copied from pudl-data-release repo
2. Dockerfile.git copies code from catalyst-cooperative github account
3. Dockerfile.local makes shallow git copy of a local repo

Notes:
Perhaps making shallow copy of local git repo is unnecessary and we can simply
copy the full state of the local repo. That may allow for testing/installing
uncommited changes into the docker image, but may also increase the size of the
resultant image.

Perhaps there is a way to make pip install work w/o the need for git files
to be also present.